ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
aiofiles==23.2.1
trimesh>=4.0.0
scipy>=1.11.0
